refactor: 사용하지 않는 Exception 삭제, 패키지 정리
