--- jdbc/src/main/java/nextstep/jdbc/exception/ExecuteUpdateException.java
@@ -1,8 +1,0 @@
-package nextstep.jdbc.exception;
-
-public class ExecuteUpdateException extends RuntimeException {
-
-    public ExecuteUpdateException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
