--- jdbc/src/main/java/nextstep/jdbc/exception/ResultSetCloseException.java
@@ -1,8 +1,0 @@
-package nextstep.jdbc.exception;
-
-public class ResultSetCloseException extends RuntimeException {
-
-    public ResultSetCloseException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
